divide packet module
